% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\alias{bw.default}
\alias{bw.list}
\alias{bw.tvlm}
\alias{bw.tvar}
\alias{bw.tvvar}
\alias{bw.tvsure}
\alias{bw.tvplm}
\alias{bw.pdata.frame}
\title{Bandwidth Selection by Cross-Validation}
\usage{
bw(x, ...)

\method{bw}{default}(
  x,
  y,
  z = NULL,
  cv.block = 0,
  est = c("lc", "ll"),
  tkernel = c("Triweight", "Epa", "Gaussian"),
  singular.ok = TRUE,
  ...
)

\method{bw}{list}(
  x,
  y,
  z = NULL,
  cv.block = 0,
  est = c("lc", "ll"),
  tkernel = c("Triweight", "Epa", "Gaussian"),
  singular.ok = TRUE,
  ...
)

\method{bw}{tvlm}(x, ...)

\method{bw}{tvar}(x, ...)

\method{bw}{tvvar}(x, ...)

\method{bw}{tvsure}(x, ...)

\method{bw}{tvplm}(x, ...)

\method{bw}{pdata.frame}(
  x,
  z = NULL,
  method,
  cv.block = 0,
  est = c("lc", "ll"),
  tkernel = c("Triweight", "Epa", "Gaussian"),
  ...
)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Other parameters passed to specific methods.}

\item{y}{A matrix or vector with the dependent variable(s).}

\item{z}{A vector with the variable over which coefficients are smooth over.}

\item{cv.block}{A positive scalar with the size of the block in leave-one block-out cross-validation.
By default 'cv.block=0' meaning leave-one-out cross-validation.}

\item{est}{The nonparametric estimation method, one of "lc" (default) for linear constant
or "ll" for local linear.}

\item{tkernel}{A character, either "Triweight" (default), "Epa" or "Gaussian" kernel function.}

\item{singular.ok}{Logical. If FALSE, a singular model is an error.}

\item{method}{A character with the choice of panel model/estimation method:
If method = \code{tvPOLS} (default) then the data is pooled estimated with time-varying OLS. 
No individual or time effects are estimated
If method = \code{tvFE} then individual effects which might be correlated with 
the regressors are estimated.
If method = \code{tvRE} then individual effects are considered random and independent
of the regressors.}
}
\value{
\code{bw} returns a vector or a scalar with the bandwith to estimate the mean or the covariance
residuals, fitted values.

A scalar or a vector of scalars.

A scalar.
}
\description{
Calculate bandwidth(s) by cross-validation for functions tvSURE, tvVAR and tvLM.
}
\examples{
##Generate data
tau <- seq(1:200)/200
beta <- data.frame(beta1 = sin(2*pi*tau), beta2 =  2*tau)
X <- data.frame(X1 = rnorm(200), X2 =  rchisq(200, df = 4))
error <- rt(200, df = 10)
y <- apply(X*beta, 1, sum) + error

##Select bandwidth by cross-validation
bw <- bw(X, y, est = "ll", tkernel = "Gaussian")

data( Kmenta, package = "systemfit" )

## x is a list of matrices containing the regressors, one matrix for each equation
x <- list()
x[[1]] <- Kmenta[, c("price", "income")]
x[[2]] <- Kmenta[, c("price", "farmPrice", "trend")]

## 'y' is a matrix with one column for each equation
y <- cbind(Kmenta$consump, Kmenta$consump)

## Select bandwidth by cross-validation
bw <- bw(x = x, y = y)

##One bandwidth per equation
print(bw)

}
