% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CEES}
\alias{CEES}
\title{Aslanidis and Casas (2013) consider a portfolio of daily US dollar exchange rates of the Australian
dollar (AUS), Swiss franc (CHF), euro (EUR), British pound (GBP), South African rand (RAND),
Brazilian real (REALB), and Japanese yen (YEN) over the period from January 1, 1999
until May 7, 2010 (T=2856 observations). This dataset contains the standarised rates after
"devolatilisation", i.e. standarising the rates using a GARCH(1,1) estimate of the volatility.}
\format{A data frame with 2856 rows and 7 variables. Below the standarised rates of daily US dollar
exchange rates of
\describe{
  \item{AUS}{Australian dollar}
  \item{CHF}{Swiss franc}
  \item{EUR}{Euro}
  \item{GBP}{British pound}
  \item{RAND}{South African rand}
  \item{REALB}{Brazilian real}
  \item{YEN}{Japanese yen}
}}
\description{
Aslanidis and Casas (2013) consider a portfolio of daily US dollar exchange rates of the Australian
dollar (AUS), Swiss franc (CHF), euro (EUR), British pound (GBP), South African rand (RAND),
Brazilian real (REALB), and Japanese yen (YEN) over the period from January 1, 1999
until May 7, 2010 (T=2856 observations). This dataset contains the standarised rates after
"devolatilisation", i.e. standarising the rates using a GARCH(1,1) estimate of the volatility.
}
\references{
Aslanidis, N. and Casas, I. (2013) Nonparametric correlation models for portfolio allocation,
\emph{Journal of Banking \& Finance}, 37, 2268 - 2283.
}
\keyword{datasets}
