% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_tvreg.R
\name{forecast}
\alias{forecast}
\title{Forecast Methods for Objects in tvReg.}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{Object of class \code{tvlm}, \code{tvar}, \code{tvvar} or \code{tvsure}.}

\item{...}{Other parameters passed to specific methods.}
}
\value{
An object of class matrix or vector with the same dimensions than  the dependent 
variable of \code{object}.
}
\description{
Forecast methods for objects with class attribute \code{tvlm}, \code{tvar}, 
\code{tvvar}, \code{tvirf} and \code{tvsure}. If the 
smoothing variable (z) in the model is non-NULL and it is a random 
variable then use function \code{predict} with parameter \code{newz}.
}
\seealso{
\code{\link{predict}}.
}
