% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tvReg.R
\name{coef.tvlm}
\alias{coef.tvlm}
\title{Coefficients and residuals of functions in tvReg}
\usage{
\method{coef}{tvlm}(object, ...)
}
\arguments{
\item{object}{An object used to select a method.}

\item{...}{Other parameters passed to specific methods.}
}
\description{
Return coefficients and residuals for objects with class attribute \code{tvlm}, 
\code{tvar}, \code{tvvar}, \code{tvirf}, \code{tvsure}.
}
