% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tvReg.R
\name{print.tvlm}
\alias{print.tvlm}
\alias{print.tvar}
\alias{print.tvsure}
\alias{print.tvvar}
\alias{print.tvirf}
\title{Print results of functions in tvReg}
\usage{
\method{print}{tvlm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{tvar}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{tvsure}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{tvvar}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{tvirf}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An x used to select a method.}

\item{digits}{An integer, indicating the minimal number of significant digits.}

\item{...}{Other parameters passed to specific methods.}
}
\description{
Print some results for objects with class attribute \code{tvlm}, \code{tvar}, \code{tvvar},
\code{tvirf}, \code{tvsure}.
}
\details{
These functions print a few results from the time-varying estimated coefficients
}
\seealso{
\code{\link{plot.tvlm}}, \code{\link{plot.tvvar}}, \code{\link{plot.tvvar}},
\code{\link{plot.tvirf}},\code{\link{plot.tvsure}}
}
