% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tvReg.R
\name{summary.tvlm}
\alias{summary.tvlm}
\alias{summary}
\alias{summary.tvar}
\alias{summary.tvsure}
\alias{summary.tvvar}
\alias{summary.tvirf}
\title{Print results of functions in tvReg}
\usage{
\method{summary}{tvlm}(object, digits = max(3, getOption("digits") - 3),
  ...)

\method{summary}{tvar}(object, digits = max(3, getOption("digits") - 3),
  ...)

\method{summary}{tvsure}(object, digits = max(3, getOption("digits") -
  3), ...)

\method{summary}{tvvar}(object, digits = max(3, getOption("digits") - 3),
  ...)

\method{summary}{tvirf}(object, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{object}{An object used to select a method.}

\item{digits}{Integer, indicating the minimal number of significant digits.}

\item{...}{Other parameters passed to specific methods.}
}
\description{
Print some results for objects with class attribute \code{tvlm}, \code{tvar}, \code{tvvar},
\code{tvirf}, \code{tvsure}.
}
\details{
These functions print a few results from the time-varying estimated coefficients
}
\seealso{
\code{\link{plot.tvlm}}, \code{\link{plot.tvvar}}, \code{\link{plot.tvvar}},
\code{\link{plot.tvirf}},\code{\link{plot.tvsure}}
}
