% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvAR.R
\name{tvAR}
\alias{tvAR}
\alias{tvar-class}
\alias{tvar}
\title{Time-varying Autoregressive model}
\usage{
tvAR(y, p = 1, bw = NULL, type = c("const", "none"), exogen = NULL,
  fixed = NULL, tkernel = "Epa", est = "lc", singular.ok = TRUE,
  ci = 0, tboot = "wild", runs = 100)
}
\arguments{
\item{y}{A vector with the dependent variable.}

\item{p}{A scalar indicating the number of lags in the model.}

\item{bw}{An opcional scalar or vector of length the number of equations. It represents
the bandwidth in the estimation of coefficients. If NULL, it is selected
by cross validation.}

\item{type}{A character 'const' if the model contains an intercept and 'none' otherwise.}

\item{exogen}{A matrix or data.frame with the exogenous variables (optional)}

\item{fixed}{(optional) numeric vector of the same length as the total number of parameters.
If supplied, only NA entries in fixed will be varied.}

\item{tkernel}{The type of kernel used in the coefficients estimation method, one of Epanesnikov ("Epa") or "Gaussian".}

\item{est}{The nonparametric estimation method, one of "lc" (default) for linear constant or "ll" for local linear.}

\item{singular.ok}{Logical. If FALSE, a singular model is an error.}

\item{ci}{A number between 0 and 1. It represents the confidence interval range, if ci = 0 (default) then confidence intervals are not
calculated. Function \code{\link{CI}} may be used at any time to calculate the CI.}

\item{tboot}{The bootstrap type, one of "wild" (default) or "wild2".}

\item{runs}{The number of bootstrap repetitiona, 100 by default.}
}
\value{
An object of class 'tvar'
The object of class \code{tvar} have the following components:
\item{tvcoef}{A vector of dimension obs (obs = number of observations - number lags),
 with the time-varying coefficients estimates.}
\item{fitted}{The fitted values.}
\item{resid}{Estimation residuals.}
\item{x}{A matrix of model data, with lagged y and exogenous variables.}
\item{y}{A vector with the original data.}
\item{bw}{Bandwidth of mean estimation.}
\item{type}{Whether the model has a constant or not.}
\item{exogen}{A matrix or data.frame with other exogenous variables.}
\item{p}{Number of lags}
\item{obs}{Number of observations in estimation.}
\item{totobs}{Number of observations in the original set.}
\item{ci}{Confidence interval range.}
\item{tboot}{Type of bootstrap.}
\item{runs}{Number of bootstrap replications.}
\item{call}{Matched call.}
}
\description{
Fits an autorregressive model with time varying coefficients
}
\examples{
## Simulate an tvAR(2) process

tt<-(1:1000)/1000
beta <- cbind( 0.5 * cos (2 * pi * tt), (tt - 0.5)^2)
y<-numeric(1000)
y[1] <- 0.5
y[2] <- -0.2

## y(t) = beta1(t) y(t-1) + beta2(t) y(t-2) + ut

for (t in 3:1000)
{
  y[t] <- y[(t-1):(t-2)] \%*\% beta[t,] + rnorm(1)
}
Y <- tail (y, 500)

## Estimate coefficients of process Y with ar.ols and tvAR
## and compare them in a plot

tvAR.2p <- tvAR(Y, p = 2, type = "none", est = "ll")
AR.2p <- ar.ols(Y, aic = FALSE, order = 2, intercept = FALSE, demean = FALSE )
plot(tail(beta[, 1], 500), ylim=range(tvAR.2p$tvcoef[, 1], tail(beta[, 1], 500)),
xlab = "", ylab = "", cex = 0.5, pch = 20)
abline(h = AR.2p$ar[1], col = 2)
lines(tvAR.2p$tvcoef[, 1], col = 4)
legend("topleft", c(expression(beta[1]),"AR", "tvAR"), col = c(1, 2, 4),
lty = 1, bty = "n")

## Estimate only coefficient from odd lags and the intercept
tvAR.6p <- tvAR(Y, p = 6, type = "const",
fixed = c(NA, 0, NA, 0, NA, 0, NA), est = "ll")

}
\seealso{
\code{\link{tvLM}} for estimation of time-varying coefficients linear models,
 and \code{\link{CI}} for confidence intervals.
}
\keyword{coefficients,}
\keyword{nonparametric}
\keyword{statistics}
\keyword{time-varying}
