% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvCov.R
\name{tvCov}
\alias{tvCov}
\title{Time varying covariance estimation}
\usage{
tvCov(x, bw, est = "lc", tkernel = "Epa")
}
\arguments{
\item{x}{A matrix.}

\item{bw}{A scalar.}

\item{est}{A character, either "lc" or "ll" for local constant or local linear.}

\item{tkernel}{A character, either "Gaussian" or "Epa" kernel types.}
}
\value{
A matrix of dimension obs x neq x neq.
}
\description{
Estimation of a time varying Sigma using the local constant or the local linear kernel
smoothing methodologies.
}
\examples{
##Generate two independent (uncorrelated series)
y <- cbind(rnorm(200, sd = 4), rnorm(200, sd = 1))

##Obtain bandwidth
bw.cov <- bwCov(y)

##Estimation variance-variance matrix
Sigma.hat <-  tvCov(y, bw = bw.cov)
##The first time estimate
print(Sigma.hat[,,1])
##The mean over time of all estimates
print(apply(Sigma.hat, 1:2, mean))
##Generate two dependent variables
y <- MASS::mvrnorm(n = 200, mu = c(0,0), Sigma = cbind(c(1, -0.5), c(-0.5, 4)))
#obtain bandwidth
bw.cov <- bwCov(y)

##Estimation variance-variance matrix
Sigma.hat <-  tvCov(y, bw = bw.cov)
##The first time estimate
print(Sigma.hat[,,1])

}
\references{
Aslanidis, N. and Casas, I (2013) Nonparametric correlation models for portfolio
allocation. \emph{Journal of Banking \& Finance}, 37, 2268-2283
}
