% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvIRF.R
\name{tvIRF}
\alias{tvIRF}
\alias{tvirf-class}
\alias{tvirf.}
\alias{tvIRF.tvvar}
\title{Time Varying Impulse Response Function}
\usage{
tvIRF(x, ...)

\method{tvIRF}{tvvar}(x, impulse = NULL, response = NULL, n.ahead = 10,
  ortho = TRUE, cumulative = FALSE, ortho.cov = c("const", "tv"),
  bw.cov = NULL, ci = 0, tboot = "wild", runs = 100, ...)
}
\arguments{
\item{x}{An object of class \code{tvvar}.}

\item{...}{Other parameters passed to specific methods.}

\item{impulse}{A character vector of the impulses, default is all variables.}

\item{response}{A character vector of the responses, default is all variables.}

\item{n.ahead}{Integer specifying the steps.}

\item{ortho}{Logical, if TRUE (the default) the orthogonalised IRF is computed.}

\item{cumulative}{Logical, if TRUE the cumulated impulse response coefficients are computed.
Default is FALSE.}

\item{ortho.cov}{A character indicating if the covariance matrix for the orthogonal tvIRF should
be estimated as a constant or time varying. Either 'const' (default) or 'tv'.}

\item{bw.cov}{A scalar (optional) with the bandwidth to estimate the errors
variance-covariance matrix. If left NULL, it is estimated.}

\item{ci}{Numeric, confidence interval range. Confidence intervals are calculated by bootstrap if
\code{ci} is different from zero. Default zero.}

\item{tboot}{Type of bootstrap, options 'wild' and 'wild2'.}

\item{runs}{An integer, specifying the runs for the bootstrap.}
}
\value{
\code{tvIRF} returns and object of class \code{tvirf} with the following components:
\item{irf}{An array dim = c(obs x number of response variables x number of impulse variables x
number steps).}
\item{Lower}{An array with the lower confidence line, if calculated.}
\item{Upper}{An array with the upper confidence line, if calculated.}
\item{response}{A character, a number of a vector with the names or positions of the response(s)
variable(s).}
\item{impulse}{A character, a number of a vector with the names or positions of the impulse(s)
variable(s).}
\item{x}{A object of class \code{tvvar}}.
\item{n.ahead}{Number of ahead impulse response functions.}
\item{ortho}{Logical, orthogonal or not impuluse response function.}
\item{cumulative}{Logical, if TRUE the cumulated impulse response coefficients are computed. Default is FALSE.}
\item{ortho.cov}{Character, either 'const' or 'tv'.}
\item{bw.cov}{A scalar with the bandwidth to estimate the errors variance-covariance
matrix.}
\item{ci}{Numeric, confidence interval range. The default is zero.}
\item{tboot}{Type of bootstrap.}
\item{runs}{Number of bootstrap replications.}
}
\description{
Computes the time varying impulse response coefficients of an object of class \code{tvvar},  obtained
with function tvVAR for n.ahead steps.
}
\examples{
## Inflation rate, unemployment rate and treasury bill interest rate for the US,
## as used by Primiceri (2005).

data(usmacro, package = "bvarsv")

##Estimate a vector autoregressive with VAR and tvVAR
model.VAR <- vars::VAR(usmacro, p = 6, type = "const")
model.tvVAR <- tvVAR(usmacro, p = 6, type = "const")

##Estimate a the impulse response functions with irf and tvIRF
model.irf <- vars::irf(model.VAR)
model.tvIRF <- tvIRF(model.tvVAR)

##Obtain 95\% confidence interval
model.tvIRF <- CI(model.tvIRF)

##plot tvIRF
plot(model.tvIRF)

##Estimate a the tvIRF with time-varying covariance function
model.tvIRF2 <- tvIRF(model.tvVAR, ortho.cov = "tv")

model.tvIRF3 <- tvIRF(model.tvVAR, ortho.cov = "tv", bw.cov = 0.2)
}
