% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvOLS.R
\name{tvOLS}
\alias{tvOLS}
\alias{tvOLS.tvlm}
\alias{tvOLS.tvar}
\alias{tvOLS.tvvar}
\title{Time-Varying Ordinary Least Squares}
\usage{
tvOLS(x, ...)

\method{tvOLS}{tvlm}(x, ...)

\method{tvOLS}{tvar}(x, ...)

\method{tvOLS}{tvvar}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{Other arguments passed to specific methods.}
}
\value{
\code{tvGLS} returns a list containing:
\item{tvcoef}{A vector of length obs, number of observations with
the time-varying estimates.}
\item{fitted}{A vector of length obs with the fited values from the estimation.}
\item{residuals}{A vector of length obs with the residuals from the estimation.}
}
\description{
\code{tvOLS} estimate time-varying coefficient of univariate 
linear models using the kernel smoothing OLS.
}
