% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvPhi.R
\name{tvPhi}
\alias{tvPhi}
\title{Time-Varying Coefficient Arrays of the MA Represention}
\usage{
tvPhi(x, nstep = 10, ...)
}
\arguments{
\item{x}{An object of class \code{tvvar}.}

\item{nstep}{An integer specifying the number of moving error coefficient matrices to be calculated.}

\item{...}{Other parameters passed to specific methods.}
}
\description{
Returns the estimated time-varying coefficient arrays of the moving average representation
of a stable \code{tvvar} object obtained with function \code{tvVAR}.
}
