% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_tvReg.R
\name{CI}
\alias{CI}
\alias{confint}
\alias{CI.default}
\alias{CI.tvirf}
\title{Deprecated function(s) in the tvReg package}
\usage{
CI(object, ...)

\method{CI}{default}(object, level = 0, runs = 0, tboot = NULL, ...)

\method{CI}{tvirf}(object, level = 0, runs = 0, tboot = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{tvsure}, class \code{tvvar} or class \code{tvirf}.}

\item{...}{Other parameters passed to specific methods.}

\item{level}{Numeric, the confidence level required (between 0 and 1).}

\item{runs}{(optional) Number of bootstrap replications.}

\item{tboot}{Type of wild bootstrap, choices 'wild'(default), 'wild2'. Option 'wild' uses the
distribution suggested by Mammen (1993) in the wild resampling, while 'wild2' uses the standard
normal.}
}
\description{
These functions are provided for compatibility with older version of
the tvReg package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{CI} \tab now a synonym for \code{\link{confint}}
}
}

