% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tvReg.R
\name{tvReg-methods}
\alias{tvReg-methods}
\alias{coef.tvlm}
\alias{coefficients.tvlm,}
\alias{coefficients.tvar,}
\alias{coefficients.tvvar,}
\alias{coefficients.tvirf,}
\alias{coefficients.tvsure}
\alias{coef.tvar}
\alias{coef.tvvar}
\alias{coef.tvirf}
\alias{coef.tvsure}
\title{Coefficients and residuals of functions in tvReg}
\usage{
\method{coef}{tvlm}(object, ...)

\method{coef}{tvar}(object, ...)

\method{coef}{tvvar}(object, ...)

\method{coef}{tvirf}(object, ...)

\method{coef}{tvsure}(object, ...)
}
\arguments{
\item{object}{An object used to select a method.}

\item{...}{Other parameters passed to specific methods.}
}
\description{
Return coefficients and residuals for objects with class attribute \code{tvlm}, 
\code{tvar}, \code{tvvar}, \code{tvirf}, \code{tvsure}.
}
