% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tvsure.control}
\alias{tvsure.control}
\title{Create list of control parameters for tvSURE}
\usage{
tvsure.control(maxiter = 1, tol = 1e-05)
}
\arguments{
\item{maxiter}{maximum number of iterations for the iterative FGLS
estimations.}

\item{tol}{tolerance level indicating when to stop the iteration for the iterative FGLS estimations}
}
\value{
A list of the above components.
}
\description{
Create a list of control pararameters for function \code{\link{tvSURE}}.
All control parameters that are not passed to this function are set to
default values.
}
\details{
If the estimation is iterative FGLS with \code{maxiter}>1, the convergence criterion is
\deqn{\sqrt{ \frac{ \sum_{i, j}
(B_{i,j,g} - B{i, j, g-1})^2 }{ \sum_{i, j} B_{i, j, g-1}^2 }} < \code{tol}}
(\eqn{B_{i, j,g}} is the ith, jth coefficient of the gth iteration step).
}
