% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.tvlm}
\alias{update.tvlm}
\alias{update.tvar}
\alias{update.tvvar}
\alias{update.tvsure}
\title{Update and Re-fit the Models of package tvReg}
\usage{
\method{update}{tvlm}(object, ...)

\method{update}{tvar}(object, ...)

\method{update}{tvvar}(object, ...)

\method{update}{tvsure}(object, ...)
}
\arguments{
\item{object}{An object of any class in package tvReg}

\item{...}{Other parameters passed to specific methods.}
}
\value{
An object of the same class than the argument *object*.
}
\description{
Update and Re-fit the Models of package tvReg
}
