% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{CI}
\alias{CI}
\alias{CI.default}
\alias{CI.tvsure}
\alias{CI.tvirf}
\title{Confidence interval calculation for objects in tvsystem}
\usage{
CI(x, ...)

\method{CI}{default}(x, ci = 0.95, runs = 100, tboot = "wild", ...)

\method{CI}{tvsure}(x, ci = 0.95, runs = 100, tboot = "wild", ...)

\method{CI}{tvirf}(x, ci = 0.95, runs = 100, tboot = "wild", ...)
}
\arguments{
\item{x}{Object of class \code{tvsure}, class \code{tvvar} or class \code{tvirf}.}

\item{...}{Other parameters passed to specific methods.}

\item{ci}{Numeric, confidence interval range (between 0 and 1). Confidence intervals are calculated by bootstrap if ci is different from zero. Default zero.}

\item{runs}{(optional) Number of bootstrap replications.}

\item{tboot}{Type of wild bootstrap, choices 'wild'(default), 'wild2'. Option 'wild' uses the
distribution suggested by Mammen (1993) in the wild resampling, while 'wild2' uses the standard
normal.}
}
\value{
an object of class \code{tvsure} with BOOT, Lower and Upper different from NULL.
}
\description{
CI is used to estimate the bootstrap confidence intervals for objects with class
attribute \code{tvlm}, \code{tvar}, \code{tvirf}, \code{tvsure}.
}
\examples{
tau <- seq(1:1000)/1000
beta <- data.frame(beta1 = sin(2*pi*tau), beta2= 2*tau)
X1 <- rnorm(1000)
X2 <- rchisq(1000, df = 4)
error <- rt(1000, df = 10)
y <- apply(cbind(X1, X2)*beta, 1, sum) + error
data <- data.frame(y = y, X1 = X1, X2 = X2)
model.tvlm <-  tvLM(y~0+X1+X2, data = data)
tvci <- CI(model.tvlm, ci = 0.95, runs = 30)
plot(tvci)


data( "Kmenta", package="systemfit" )
eqDemand <- consump ~ price + income
eqSupply <- consump ~ price + farmPrice + trend
system <- list( demand = eqDemand, supply = eqSupply )

##Because parameter 'ci'=0 by default, no confidence interval is calculated
tvfgls1.fit <- tvSURE(system, data = Kmenta, method="tvFGLS")

##Calculate 95\% confidence interval of our object using a resampling of size 100
tvfgls1.fit <- CI (tvfgls1.fit, ci = 0.95, runs = 100)

##Once the first confidence interval is calculated, the same resamples are used for other
##confidence intervals, which makes the process faster for consecutives calls of function CI.
tvfgls1.fit <- CI (tvfgls1.fit, ci = 0.90)
}
\references{
Chen, X. B., Gao, J., Li, D., and Silvapulle, P (2013) Nonparametric estimation and
parametric calibration of time-varying coefficient realized volatility models.
\emph{Monash Econometrics and Business StatisticsWorking Papers 21/13}
URL:https://ideas.repec.org/p/msh/ebswps/2013-21.html.

Mammen, E (1993) Bootstrap and wild bootstrap for high dimensional linear models,
\emph{ Annals of Statistics}, 21, 255-285.
}
