% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tvsystem.R
\name{plot.tvsure}
\alias{plot.tvsure}
\alias{plot.tvlm}
\alias{plot.tvar}
\alias{plot.tvvar}
\alias{plot.tvirf}
\title{Plot Methods for objects in tvReg}
\usage{
\method{plot}{tvsure}(x, eqs = NULL, vars = NULL, ...)

\method{plot}{tvlm}(x, ...)

\method{plot}{tvar}(x, ...)

\method{plot}{tvvar}(x, ...)

\method{plot}{tvirf}(x, obs.index = NULL, impulse = NULL, response = NULL,
  plot.type = c("multiple", "single"), names = NULL, main = NULL,
  sub = NULL, ylab = NULL, xlab = NULL, nc, mar.multi = c(0, 4, 0, 4),
  oma.multi = c(6, 4, 6, 4), adj.mtext = NA, padj.mtext = NA, ...)
}
\arguments{
\item{x}{An x used to select a method.}

\item{eqs}{Character vector (optional) with the equation(s) number(s) or
equation name(s) of the coefficients to be plotted.}

\item{vars}{Character vector (optional) with the variable number(s) or
variable name(s) of the coefficients to be plotted.}

\item{...}{Other parameters passed to specific methods.}

\item{obs.index}{Scalar (optional), the time at which the impulse response is plotted.
If left NULL, the mean over the whole period is plotted (this values should be similar to
the estimation using a non time-varying VAR method).}

\item{impulse}{Character  vector (optional) of the impulses, default is all variables.}

\item{response}{Character vector (optional) of the responses, default is all variables.}

\item{plot.type}{Character, if multiple all plots are drawn in a single device,
otherwise the plots are shown consecutively.}

\item{names}{Character vector (optional), the variables names to be plotted.
If left NULL, all variables are plotted.}

\item{main}{Character vector, the titles of the plot.}

\item{sub}{Character, sub title in plot.}

\item{ylab}{Character vector signifying the labels for the y-axis.}

\item{xlab}{Character vector signifying the labels for the x-axis.}

\item{nc}{Integer, number of columns for multiple plot.}

\item{mar.multi}{Setting of margins, if plot.type = "multiple".}

\item{oma.multi}{Setting of margins, if plot.type = "multiple".}

\item{adj.mtext}{Adjustment for mtext(), only applicable if plot.type = "multiple".}

\item{padj.mtext}{Adjustment for mtext(), only applicable if plot.type = "multiple".}
}
\description{
Plot methods for objects with class attribute \code{tvlm}, \code{tvar}, \code{tvvar},
\code{tvirf}, \code{tvsure}.
}
\examples{

data( "Kmenta", package = "systemfit" )
eqDemand <- consump ~ price + income
eqSupply <- consump ~ price + farmPrice + trend
system <- list( demand = eqDemand, supply = eqSupply )

## tvOLS estimation
tvols.fit <- tvSURE(system, data = Kmenta)
## 95\% confidence interval using Mammen's wild bootstrap.
tvols.fit <- CI(tvols.fit, level = 0.95)
plot(tvols.fit)

## Time-varying Five Factor Model
data(FF5F)
## SMALL/LoBM porfolios time-varying three factor model
eqNA<-NA.SMALL.LoBM ~ NA.Mkt.RF + NA.SMB+ NA.HML
eqJP<-JP.SMALL.LoBM ~ JP.Mkt.RF + JP.SMB+ JP.HML
eqAP<-AP.SMALL.LoBM ~ AP.Mkt.RF + AP.SMB+ AP.HML
eqEU<-EU.SMALL.LoBM ~ EU.Mkt.RF + EU.SMB+ EU.HML
system<-list(NorthA = eqNA, JP = eqJP, AP = eqAP, EU = eqEU)

## Fit a time-varying coefficients SURE model
ff5f.tv<-tvSURE(system, data = FF5F, method = "tvFGLS", est = "ll",
bw = c(0.11, 0.43, 0.49, 0.27))

## 95\% confidence interval using Mammen's wild bootstrap.
ff5f.tv <- CI(ff5f.tv, level = 0.95, runs = 30)

## Plot the intercepts
plot(ff5f.tv, vars = 1)

## Inflation rate, unemployment rate and treasury bill interest rate for the US,
## as used by Primiceri (2005).

data(usmacro, package = "bvarsv")

##Estimate a time-varying coefficients vector autoregressive and its
##impulse response function
model.tvVAR <- tvVAR(usmacro, p = 6, type = "const")
model.tvIRF <- tvIRF(model.tvVAR)

## Plot residuals and fitted values of tvVAR

##Obtain 95\% confidence interval of the impulse response function
model.tvIRF <- CI(model.tvIRF)

##plot the mean tvIRF over the whole period
plot(model.tvIRF)

##plot the tvIRF at time 50
plot(model.tvIRF, obs.index = 50)

##plot the effect of a shock in inflation on unemployment
plot(model.tvIRF, impulse = "inf", response = "une")

}
\seealso{
\code{\link{tvLM}}, \code{\link{tvAR}}, \code{\link{tvVAR}},
\code{\link{tvSURE}}
}
