% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.tvsure}
\alias{update.tvsure}
\alias{update.tvlm}
\alias{update.tvar}
\alias{update.tvvar}
\title{Update and Re-fit the Models of package tvReg}
\usage{
\method{update}{tvsure}(object, y = NULL, ...)

\method{update}{tvlm}(object, y = NULL, ...)

\method{update}{tvar}(object, y = NULL, ...)

\method{update}{tvvar}(object, y = NULL, ...)
}
\arguments{
\item{object}{An object used to select a method.}

\item{y}{The dependent variable to update the model.}

\item{...}{Other parameters passed to specific methods.}
}
\value{
An object of class \code{tvsure}.

An object of class \code{tvlm}.

An object of class \code{tvar}.

An object of class \code{tvvar}.
}
\description{
Update and Re-fit the Models of package tvReg
}
